change the delete button confirm to destructive and import employee types
